requirements.txt
# General requirements
matplotlib==3.7.*
numpy==1.24.*
pydantic==1.10.*
pyyaml==6.0
torch==2.0.0
torchvision==0.15.*
tqdm==4.65.*
